Clouds partly optimised; circleMorphing now numbers nodes.
Node numbering reveals that vector interpolating working, but that
triangle is plotted incorrectly — nodes just bustle around each corner
point. So, we need to plot this out differently.

Clouds not really optimised, but every 5 seconds or so we are
repositioning them left and right relative to boo’s x. So, we want to
be able to test that we only reposition clouds that are out of view,
and that we reposition them out of view too, so as to create illusion
of consistency.